daemon/images: move SquashImage to its own file
